Small currentDirection Fix
Added mushrooms to the list of walkables for the people who use the default S shaped farm layout with bigger mushroom hitboxes to farm mushroom. The mushrooms would occasionally spread to the dirt floor, causing the player to change direction when colliding with it.

Although this could easily be avoided, I decided to add a fix for this anyway since it is harmless.