Greenhouse & Prediction Anti-Cheat Update
- added support for new crops and pests after The Greenhouse Update, thanks Steinebeisser!
- updated movement logic due to prediction anti-cheat update

As for the Discord server... There are currently no plans to continue the development.
Our Discord vanity was stolen by a ratter, and despite hundreds of reports and multiple DMCAs, Discord took no action.
Jelly is basically dead (as he's focused on IRL stuff) meaning Mighty Miner is dead as well.
Farm Helper needs a full rewrite or at least extensive bug fixes, but there are little to no contributors willing to work on it.
May2Bee and I (yuro) also don't have much time to continue development since we're focused on other projects such as Nebula Client.
Unfortunately, this is the downfall of the Jelly community. We're genuinely grateful for the time we've spent with all of you guys.
All good things come to an end eventually.